Day 24. Refactoring and debugging. operation create/edit components refactored - code duplication eliminated.
